use identical reqwest client for rest and opendal
